basicGetFunctions
