--- Backend/getToken.js
@@ -1,4 +1,0 @@
-function getToken() {
-    const token = "BQBrZwvZwpRWoH822GOn7427JLSKd9p0_1QRdNf6wqmv-F-KbQ-i5jYhZZvCQe5u-0y3OA_CMb2YpnTSahahpmnuvVdeQqkpvGNnw3iu_REhTNYK9RP4iMY7GuKE1BZhm4BMzzlfgGPaRYV2hIPjiCL-RuPMguVeir9NWDDSBK6_DfLo4WHMjyMY8TzoWGPLy6TO8VX3JAPgg1tyefPybL1_0JKpZN0ZS-OYqc6PMFmS3aEREF-0gH_16MW1qRmzecsqJ7sX55i06czKTn1XCT7yCCz6SjRXzPA9hfWaYRAUGYJjEoYCJHtscENVy1fVI1oasoTm07420JvYVgQ3BGuwoA";
-return token;
-}
